requirements.txt
langchain
langchain_openai
langchain_community
langchain_huggingface
huggingface_hub
python-dotenv
streamlit